+++ TrashBin/faster_whisper/__init__.py
@@ -1,0 +1,13 @@
+from faster_whisper.audio import decode_audio
+from faster_whisper.transcribe import WhisperModel
+from faster_whisper.utils import available_models, download_model, format_timestamp
+from faster_whisper.version import __version__
+
+__all__ = [
+    "available_models",
+    "decode_audio",
+    "WhisperModel",
+    "download_model",
+    "format_timestamp",
+    "__version__",
+]
