+++ TrashBin/faster_whisper/version.py
@@ -1,0 +1,3 @@
+"""Version information."""
+
+__version__ = "0.10.0"
