--- faster_whisper/version.py
@@ -1,3 +1,0 @@
-"""Version information."""
-
-__version__ = "0.10.0"
