+++ requirements.txt
@@ -6,7 +6,9 @@ anyio==4.2.0
 attrs==23.1.0
 av==10.0.0
 backoff==2.2.1
+beautifulsoup4==4.12.3
 blinker==1.7.0
+bs4==0.0.2
 cachetools==5.3.2
 certifi==2023.11.17
 charset-normalizer==3.3.2
@@ -41,16 +43,34 @@ langchain==0.0.352
 langchain-community==0.0.6
 langchain-core==0.1.3
 langsmith==0.0.75
+llvmlite==0.42.0
 markdown-it-py==3.0.0
 MarkupSafe==2.1.3
 marshmallow==3.20.1
 mdurl==0.1.2
 monotonic==1.6
+more-itertools==10.2.0
 mpmath==1.3.0
 multidict==6.0.4
 mypy-extensions==1.0.0
+networkx==3.2.1
+numba==0.59.1
 numpy==1.26.2
+nvidia-cublas-cu12==12.1.3.1
+nvidia-cuda-cupti-cu12==12.1.105
+nvidia-cuda-nvrtc-cu12==12.1.105
+nvidia-cuda-runtime-cu12==12.1.105
+nvidia-cudnn-cu12==8.9.2.26
+nvidia-cufft-cu12==11.0.2.54
+nvidia-curand-cu12==10.3.2.106
+nvidia-cusolver-cu12==11.4.5.107
+nvidia-cusparse-cu12==12.1.0.106
+nvidia-nccl-cu12==2.19.3
+nvidia-nvjitlink-cu12==12.4.99
+nvidia-nvtx-cu12==12.1.105
 onnxruntime==1.16.3
+openai-whisper==20231117
+opencv-python==4.9.0.80
 overrides==7.4.0
 packaging==23.2
 pandas==2.1.4
@@ -74,28 +94,36 @@ python-dotenv==1.0.0
 pytz==2023.3.post1
 PyYAML==6.0.1
 referencing==0.32.0
+regex==2023.12.25
 requests==2.31.0
 rich==13.7.0
 rpds-py==0.15.2
+safetensors==0.4.2
 semver==3.0.2
 six==1.16.0
 smmap==5.0.1
 sniffio==1.3.0
+soupsieve==2.5
 SQLAlchemy==2.0.23
 starlette==0.27.0
 streamlit==1.29.0
+streamlit-option-menu==0.3.12
 sympy==1.12
 tenacity==8.2.3
+tiktoken==0.6.0
 tokenizers==0.15.0
 toml==0.10.2
 toolz==0.12.0
+torch==2.2.2
 tornado==6.4
 tqdm==4.66.1
+transformers==4.39.2
+triton==2.2.0
 typing-inspect==0.9.0
 typing_extensions==4.9.0
 tzdata==2023.3
 tzlocal==5.2
-urllib3==2.1.0
+urllib3==2.2.1
 uvicorn==0.25.0
 uvloop==0.19.0
 validators==0.22.0
